refactor(a1): extract the movers engine to core/movers.py (BR-CODE-1)
Heavy-helper pass, final slice — moves the largest remaining app.py chunk, the
self-contained storefront "movers" engine, into a new core/movers.py:
  - _compute_movers (the entry; exported as compute_movers)
  - the 5 _section_* rail builders (moving_fast/price_drops/climbing/specials/featured)
  - 3 classifiers: _classify_marquee_competition, _has_canadian_team,
    _classify_holiday_proximity (+ constants _MARQUEE_PATTERNS / _CANADIAN_TEAM_RE
    / _HOLIDAY_PRIORITY)

The whole cluster was movers-exclusive (verified: every member ref-count = def +
its in-cluster call) and depends outward only on core.helpers + core.broker_helpers,
so the move is verbatim. app.py imports compute_movers aliased to the historical
_compute_movers; the two store/broker movers call sites are unchanged.

tests/test_store_home.py repointed: the section builders + holiday-proximity
classifier are now reached via `core.movers` (app still re-exports _compute_movers
/ _classify_world_cup / _or_ilike_clause, so those references are untouched). Its
63 tests are the end-to-end regression guard and pass against core.movers.

app.py 7709 -> 6965 LOC (-744, the largest single reduction). Full suite 369
passed; the 14 remaining local failures are env-only deps (supabase.Client,
pyo3 crypto) in untouched modules, which CI provides.
